+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes in keycloak-group-management will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2025-02-05
 
 ### Added 
 - Admins can directly add new members to groups
+
+### Fixed
+- Fix problem deleting group with aup in configuration
 
 ## [1.4.1]- 2025-02-03